// src/types/soap.ts
export type IFolderView =
	| 'search folder'
	| 'tag'
	| 'conversation'
	| 'message'
	| 'contact'
	| 'document'
	| 'appointment'
	| 'virtual conversation'
	| 'remote folder'
	| 'wiki'
	| 'task'
	| 'chat';

type mapContactIds = {
	ids: string;
};

export type ISoapFolderObj = {
	absFolderPath: string;
	activesyncdisabled: boolean;
	deletable: boolean;
	folder?: Array<ISoapFolderObj>;
	i4ms: number;
	i4next: number;
	id: string;
	color: string;
	rgb: string;
	/** Parent ID */ l: string;
	luuid: string;
	ms: number;
	/** Count of non-folder-panel items */ n: number;
	name: string;
	rev: number;
	/** Size */ s: number;
	/** Count of unread messages */ u?: number;
	uuid: string;
	absParent?: string;
	view: IFolderView;
	cn: Array<mapContactIds>;
	webOfflineSyncDays: number;
};